+++ packages/grid/src/features/standard-table/components/StandardTable.tsx
@@ -1,7 +1,14 @@
 import { Box, useDomId } from "@stenajs-webui/core";
+import cx from "classnames";
 import * as React from "react";
 import { useMemo } from "react";
 import { StandardTableConfig } from "../config/StandardTableConfig";
+import { GroupConfigsForRowsContext } from "../context/GroupConfigsForRowsContext";
+import { OnKeyDownContext } from "../context/OnKeyDownContext";
+import {
+  StandardTableColumnGroupOrderContext,
+  StandardTableUsingColumnGroupsContext,
+} from "../context/StandardTableColumnOrderContext";
 import {
   StandardTableActionsContext,
   StandardTableConfigContext,
@@ -10,20 +17,17 @@ import {
   StandardTableTableIdContext,
   TableContext,
 } from "../context/StandardTableStateContext";
+import { StandardTableVariantContext } from "../context/StandardTableVariantContext";
+import { createColumnConfigsForRows } from "../features/column-groups/ColumnGroupFactory";
+import { ColumnGroupRow } from "../features/column-groups/ColumnGroupRow";
+import { calculateColumnIndexPerColumnId } from "../features/column-index-per-column-id/ColumnIndexCalculator";
+import { ColumnIndexPerColumnIdContext } from "../features/column-index-per-column-id/ColumnIndexPerColumnIdContext";
 import { useLocalStateTableContext } from "../hooks/UseLocalStateTableContext";
+import { createStandardTableInitialState } from "../redux/StandardTableReducer";
+import { StandardTableOnKeyDown } from "../types/StandardTableOnKeyDown";
+import styles from "./StandardTable.module.css";
 import { StandardTableContent } from "./StandardTableContent";
 import { StandardTableHeadRow } from "./StandardTableHeadRow";
-import { createStandardTableInitialState } from "../redux/StandardTableReducer";
-import styles from "./StandardTable.module.css";
-import cx from "classnames";
-import { StandardTableVariantContext } from "../context/StandardTableVariantContext";
-import {
-  StandardTableColumnGroupOrderContext,
-  StandardTableUsingColumnGroupsContext,
-} from "../context/StandardTableColumnOrderContext";
-import { ColumnGroupRow } from "./column-groups/ColumnGroupRow";
-import { createColumnConfigsForRows } from "../util/ColumnGroupFactory";
-import { GroupConfigsForRowsContext } from "../context/GroupConfigsForRowsContext";
 
 export interface StandardTableProps<
   TItem,
@@ -92,6 +96,12 @@ export interface StandardTableProps<
    * The order of column groups. If set, it overrides the order set in the config.
    */
   columnGroupOrder?: Array<TColumnGroupKey>;
+
+  /**
+   * onKeyDown for the table. First argument is the event, second argument is
+   * an object that contains columnId and item for the focused cell.
+   */
+  onKeyDown?: StandardTableOnKeyDown<TItem, TColumnKey>;
 }
 
 export type StandardTableVariant =
@@ -111,6 +121,7 @@ export const StandardTable = function StandardTable<
   columnGroupOrder,
   tableId,
   variant = "standard",
+  onKeyDown,
   ...props
 }: StandardTableProps<TItem, TColumnKey, TColumnGroupKey>) {
   const generatedTableId = useDomId();
@@ -146,6 +157,11 @@ export const StandardTable = function StandardTable<
     [config.columnGroups, config.columnGroupOrder, config.columnOrder]
   );
 
+  const columnIndexPerColumnId = useMemo(
+    () => calculateColumnIndexPerColumnId(config),
+    [groupConfigsForRows]
+  );
+
   return (
     <Box className={cx(styles.standardTable, styles[variant])}>
       <StandardTableVariantContext.Provider value={variant}>
@@ -158,22 +174,32 @@ export const StandardTable = function StandardTable<
                 <GroupConfigsForRowsContext.Provider
                   value={groupConfigsForRows}
                 >
-                  <StandardTableUsingColumnGroupsContext.Provider
-                    value={usingColumnGroups}
+                  <ColumnIndexPerColumnIdContext.Provider
+                    value={columnIndexPerColumnId}
                   >
-                    <StandardTableColumnGroupOrderContext.Provider
-                      value={columnGroupOrder ?? config.columnGroupOrder ?? []}
+                    <StandardTableUsingColumnGroupsContext.Provider
+                      value={usingColumnGroups}
                     >
-                      {(columnGroupOrder || config.columnGroupOrder) && (
-                        <ColumnGroupRow height={"var(--current-row-height)"} />
-                      )}
-                      <StandardTableHeadRow
-                        items={props.items}
-                        height={"var(--current-row-height)"}
-                      />
-                      <StandardTableContent variant={variant} {...props} />
-                    </StandardTableColumnGroupOrderContext.Provider>
-                  </StandardTableUsingColumnGroupsContext.Provider>
+                      <StandardTableColumnGroupOrderContext.Provider
+                        value={
+                          columnGroupOrder ?? config.columnGroupOrder ?? []
+                        }
+                      >
+                        <OnKeyDownContext.Provider value={onKeyDown}>
+                          {(columnGroupOrder || config.columnGroupOrder) && (
+                            <ColumnGroupRow
+                              height={"var(--current-row-height)"}
+                            />
+                          )}
+                          <StandardTableHeadRow
+                            items={props.items}
+                            height={"var(--current-row-height)"}
+                          />
+                          <StandardTableContent variant={variant} {...props} />
+                        </OnKeyDownContext.Provider>
+                      </StandardTableColumnGroupOrderContext.Provider>
+                    </StandardTableUsingColumnGroupsContext.Provider>
+                  </ColumnIndexPerColumnIdContext.Provider>
                 </GroupConfigsForRowsContext.Provider>
               </StandardTableConfigContext.Provider>
             </StandardTableActionsContext.Provider>