+++ packages/grid/src/features/standard-table/components/StandardTableCell.tsx
@@ -1,7 +1,10 @@
 import * as React from "react";
-import { useMemo } from "react";
+import { KeyboardEventHandler, useCallback, useMemo } from "react";
+import { useGridCell } from "../../grid-cell/hooks/UseGridCell";
+import { useOnKeyDownContext } from "../context/OnKeyDownContext";
+import { useColumnIndexPerColumnIdContext } from "../features/column-index-per-column-id/ColumnIndexPerColumnIdContext";
 import { useCellBackgroundByColumnId } from "../hooks/UseCellBackground";
-import { useColumnFromConfig } from "../hooks/UseColumnFromConfig";
+import { useColumnConfigById } from "../hooks/UseColumnConfigById";
 import {
   useStandardTableConfig,
   useStandardTableId,
@@ -9,7 +12,6 @@ import {
 import { formatValueLabel } from "../util/LabelFormatter";
 import { StandardTableCellUi } from "./StandardTableCellUi";
 import { TextCell } from "./TextCell";
-import { useGridCell } from "../../grid-cell/hooks/UseGridCell";
 
 export interface StandardTableCellProps<TItem> {
   columnId: string;
@@ -29,12 +31,12 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
   disableBorderLeft,
 }: StandardTableCellProps<TItem>) {
   const {
-    columnOrder,
     enableGridCell,
-    showRowCheckbox,
     gridCellOptions: gridCellOptionsForTable,
   } = useStandardTableConfig();
   const tableId = useStandardTableId();
+  const onKeyDown = useOnKeyDownContext();
+  const { numNavigableColumns } = useColumnIndexPerColumnIdContext();
 
   const {
     itemValueResolver,
@@ -52,7 +54,7 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
     sticky,
     zIndex,
     left,
-  } = useColumnFromConfig(columnId);
+  } = useColumnConfigById(columnId);
 
   const itemValue = useMemo(() => {
     if (itemValueResolver) {
@@ -79,11 +81,18 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
       ? isEditable(item)
       : undefined;
 
+  const onKeyDownHandler = useCallback<KeyboardEventHandler<HTMLDivElement>>(
+    (ev) => {
+      onKeyDown?.(ev, { columnId, item });
+    },
+    [onKeyDown]
+  );
+
   const gridCell = useGridCell<string>(label, {
     colIndex,
     rowIndex,
     numRows,
-    numCols: columnOrder?.length ?? 0 + (showRowCheckbox ? 1 : 0),
+    numCols: numNavigableColumns,
     tableId,
     isEditable: editable,
     onChange: onChange
@@ -120,6 +129,7 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
       zIndex={zIndex}
       left={left}
       shadow={sticky ? "var(--swui-sticky-column-shadow-right)" : undefined}
+      onKeyDown={onKeyDownHandler}
     >
       {content}
     </StandardTableCellUi>