+++ packages/grid/src/features/standard-table/context/OnKeyDownContext.ts
@@ -1,0 +1,9 @@
+import { createContext, useContext } from "react";
+import { StandardTableOnKeyDown } from "../types/StandardTableOnKeyDown";
+
+export const OnKeyDownContext = createContext<
+  StandardTableOnKeyDown<any, any> | undefined
+>(undefined);
+
+export const useOnKeyDownContext = <TItem, TColumnKey extends string>() =>
+  useContext(OnKeyDownContext) as StandardTableOnKeyDown<TItem, TColumnKey>;