+++ packages/grid/src/features/standard-table/helpers/cell-renderers/editable-text-cell/EditableTextCell.tsx
@@ -1,7 +1,7 @@
 import { Indent, Text } from "@stenajs-webui/core";
 import { TextInput } from "@stenajs-webui/forms";
 import * as React from "react";
-import { StandardTableCellRenderer } from "../../../config/StandardTableConfig";
+import { StandardTableCellRenderer } from "../../../config/StandardTableColumnConfig";
 
 export const createStandardEditableTextCell = <
   TItemValue,