+++ packages/grid/src/features/standard-table/types/StandardTableOnKeyDown.ts
@@ -1,0 +1,7 @@
+import * as React from "react";
+import { StandardTableOnKeyDownArgs } from "../config/StandardTableConfig";
+
+export type StandardTableOnKeyDown<TItem, TColumnKey extends string> = (
+  ev: React.KeyboardEvent<HTMLDivElement>,
+  args: StandardTableOnKeyDownArgs<TItem, TColumnKey>
+) => void;