+++ packages/grid/src/features/table-ui/components/table/TableRow.tsx
@@ -4,12 +4,22 @@ import {
   defaultTableHeadRowHeight,
   tableBorder,
 } from "../../../../config/TableConfig";
+import styles from "./TableRow.module.css";
+import cx from "classnames";
 
 interface Props extends BoxProps {}
 
 export const TableRow: React.FC<Props> = ({
   height = defaultTableHeadRowHeight,
+  className,
   ...boxProps
 }) => {
-  return <Row borderTop={tableBorder} height={height} {...boxProps} />;
+  return (
+    <Row
+      borderTop={tableBorder}
+      height={height}
+      className={cx(styles.tableRow, className)}
+      {...boxProps}
+    />
+  );
 };